+++ lib/glimmer/swt/text_proxy.rb
@@ -72,7 +72,7 @@ module Glimmer
                 doit
               }
             }
-          }
+          },
         }
       end