+++ lib/glimmer/swt/widget_proxy.rb
@@ -152,8 +152,8 @@ module Glimmer
       end
       
       def remove_all_listeners
-        observation_request_to_event_mapping.keys.each do |keyword|
-          [observation_request_to_event_mapping[keyword]].flatten.each do |mapping|
+        effective_observation_request_to_event_mapping.keys.each do |keyword|
+          [effective_observation_request_to_event_mapping[keyword]].flatten.each do |mapping|
             observation_requests[keyword].to_a.each do |event_listener|
               event = mapping[:event]
               event_handler = mapping[:event_handler]
@@ -268,6 +268,21 @@ module Glimmer
         {}
       end
       
+      def effective_observation_request_to_event_mapping
+        default_observation_request_to_event_mapping.merge(observation_request_to_event_mapping)
+      end
+      
+      def default_observation_request_to_event_mapping
+        {
+          'on_focus_gained' => {
+            event: 'focus',
+          },
+          'on_focus_lost' => {
+            event: 'blur',
+          },
+        }
+      end
+      
       def name
         self.class.name.split('::').last.underscore.sub(/_proxy$/, '').gsub('_', '-')
       end
@@ -357,10 +372,10 @@ module Glimmer
       end
       
       def handle_observation_request(keyword, &event_listener)
-        return unless observation_request_to_event_mapping.keys.include?(keyword)
+        return unless effective_observation_request_to_event_mapping.keys.include?(keyword)
         event = nil
         delegate = nil
-        [observation_request_to_event_mapping[keyword]].flatten.each do |mapping|
+        [effective_observation_request_to_event_mapping[keyword]].flatten.each do |mapping|
           observation_requests[keyword] ||= Set.new
           observation_requests[keyword] << event_listener
           event = mapping[:event]