+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.2
 
 - Fix issue with table item selection for booking not working after editing has been added
+- Fix escape keyboard event handling for combo table editor
 
 ## 0.7.1