+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - `date` `table_column` `editor`
 - `time` `table_column` `editor`
 - Fix issue with table item selection for booking not working after editing has been added
+- Implement `on_focus_gained`, `on_focus_lost` universally on all widgets
+- Fix issue with hello_table button/combo not being centered (yet stretched)
 - Fix escape keyboard event handling for combo table editor
 
 ## 0.7.1